+++ changelog/5353.added.md
@@ -1,6 +1,8 @@
 - Added `ElevenLabsDialogueTTSService`, a WebSocket TTS service for ElevenLabs Eleven v3 models (`eleven_v3` and `eleven_v3_conversational`), which `ElevenLabsTTSService` can't reach. It needs workspace access to ElevenLabs' Text-to-Dialogue API. `stability` is the only voice setting Text-to-Dialogue reads, and text is always aggregated into sentences:
 
   ```python
+  from pipecat.services.elevenlabs.dialogue.tts import ElevenLabsDialogueTTSService
+
   tts = ElevenLabsDialogueTTSService(
       api_key=os.getenv("ELEVENLABS_API_KEY"),
       settings=ElevenLabsDialogueTTSService.Settings(
